refactor(events): move the event bus to brain/platform and drop the vault→cortex import (#728)
`publish_safe` lived in brain/systems/cortex/events.py, so the vault reached
into Cortex through a function-local import to publish one grant-prompt event.
The module was never a Cortex concern: its only module-level brain imports are
brain.platform.db.*.

- Move brain/systems/cortex/events.py -> brain/platform/events.py. No shim: the
  old module is gone and all 27 importers point at the new path.
- Remove the module's single brain/systems/** edge. `publish()` used a lazy
  `from brain.systems.runs.event_log import async_record_run_event`; the caller
  now passes the recorder through `run_event_scope(..., recorder=...)`. A
  module-global registry was rejected: set_publisher is registered only in
  brain/app/api/main.py, and the deployment runs four processes.
- brain/systems/vault/agent_access.py imports publish_safe at module level —
  the deferral only existed to dodge the cortex cycle, which no longer exists.
- Drop the architecture-boundary allowlist entry for
  brain/platform/browser/service.py -> brain.systems.cortex.events; that import
  is now platform->platform.

Closes #728
