+++ brain/systems/cortex/resources/leases.py
@@ -11,7 +11,7 @@ from sqlalchemy import select
 
 from brain.kernel.common.time import utcnow as _shared_utcnow
 
-from brain.systems.cortex.events import publish_safe
+from brain.platform.events import publish_safe
 from brain.platform.db.models.resource_pool import ResourceLease
 from brain.platform.db.repositories.unit_of_work import UnitOfWork
