+++ tests/fixtures/architecture-boundary-allowlist.json
@@ -74,16 +74,6 @@
     },
     {
       "source": "brain/platform/browser/service.py",
-      "import": "brain.systems.cortex.events",
-      "source_layer": "brain.platform",
-      "target_layer": "brain.systems",
-      "allowed_occurrences": 1,
-      "owner": "architecture-boundaries",
-      "planned_removal": "Extract the shared behavior behind the intended lower-layer contract, then remove or lower this allowance.",
-      "reason": "Platform code calls system behavior; replace with callback, event, or interface owned outside the adapter."
-    },
-    {
-      "source": "brain/platform/browser/service.py",
       "import": "brain.systems.cortex.resources.telemetry",
       "source_layer": "brain.platform",
       "target_layer": "brain.systems",
